indented styles and moved formatNum to utils for multiple case uses
